// src/app/service/interfaces/genres.interface.ts
export interface MovieGenres {
  actionMovieResult: [];
  animationMovieResult: [];
  bollywoodMovieResult: [];
  comedyMovieResult: [];
  dramaMovieResult: [];
  documentaryMovieResult: [];
  fantasyMovieResult: [];
  horrorMovieResult: [];
  internationalMovieResult: [];
  mystryMovieResult: [];
  romanceMovieResult: [];
  scifiMovieResult: [];
  thrillerMovieResult: [];
  teenDramaMovieResult: [];
}

export interface TvGenres {
  actionTvResult: [];
  animationTvResult: [];
  bollywoodTvResult: [];
  comedyTvResult: [];
  dramaTvResult: [];
  documentaryTvResult: [];
  fantasyTvResult: [];
  horrorTvResult: [];
  internationalTvResult: [];
  mystryTvResult: [];
  romanceTvResult: [];
  scifiTvResult: [];
  thrillerTvResult: [];
  teenDramaTvResult: [];
}